second commit
it is a EOG signal processing base on its images .
I first transform every 30 second  of signal with step size 05 second to picture. Then in phase 2  file, I used 3 layers of CNN and 3 layers of fully connected  for classifying of it.
Unfortunately i could not improve the result in this time.